アドミン基本的な機能を実装

+++ miraiSugoroku/src/main/java/ksp/group3/miraiSugoroku/entity/Square.java
@@ -20,7 +20,7 @@ public class Square {
     String title;
     Long squareEventId;
     Long eventId;
-    Long groupId;
+    int groupId;
     String description;
     Long creatorId;
     int squareEffect; // マスの効果が良い（1）か悪い（2）かを返す
